% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlackJack.R
\name{BlackJack}
\alias{BlackJack}
\title{Draw 3 cards: BlackJack}
\usage{
BlackJack(n = 30)
}
\arguments{
\item{n}{draws (integer value, default = 30)}
}
\value{
sum of card values for each draw
}
\description{
Draw 3 cards: BlackJack
}
