% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProbDistributions.R
\name{chisq_area}
\alias{chisq_area}
\title{Adapted from: https://r-coder.com/normal-distribution-r
Plot Chi-Square distribution}
\usage{
chisq_area(df = 5, lb, ub, acolor = "lightgray", ...)
}
\arguments{
\item{df}{Degrees of freedom (default = 5)}

\item{lb:}{lower bound of the area}

\item{ub:}{upper bound of the area}

\item{acolor:}{color of the area
...: additional arguments to be passed to lines function}
}
\description{
Adapted from: https://r-coder.com/normal-distribution-r
Plot Chi-Square distribution
}
