% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoinToss.R
\name{CoinToss}
\alias{CoinToss}
\title{Tossing a coin}
\usage{
CoinToss(n = 1, p = 0.5)
}
\arguments{
\item{n}{number of coins to toss (integer value, default = 1)}

\item{p}{probability of a 'success' (Tail/Head), default = 0.5}
}
\value{
either 'Pile' or 'Face'
}
\description{
Tossing a coin
}
