% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConDiscord.R
\name{ConDiscord}
\alias{ConDiscord}
\title{Calcul du nombre de concordances/discordances/égalités
dans un tableau de contingence (2 variables ordinales)}
\usage{
ConDiscord(x)
}
\arguments{
\item{x}{Tableau de contingence bidimensionnel
x est un tableau (matrice, array, table ou data frame) bidimensionnel}
}
\description{
Calcul du nombre de concordances/discordances/égalités
dans un tableau de contingence (2 variables ordinales)
}
\examples{
x <- table(sample(1:3, 50, replace=TRUE), sample(1:5, 50, replace=TRUE) )
ConDiscord(x)
}
