% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfInt.R
\name{ConfInt}
\alias{ConfInt}
\title{Confidence interval of the mean - Simulation
Samples drawn from a Normal distribution}
\usage{
ConfInt(nexp = 100, nrep = 5000, n = 25, mu = 50, stdev = 8, nconf = 0.95)
}
\arguments{
\item{nexp}{= Number of samples to draw}

\item{nrep}{= Number of resamplings}

\item{n}{= sample size}

\item{mu}{= Population mean}

\item{stdev}{= Population standard deviation}

\item{nconf}{= Confidence level}
}
\description{
Confidence interval of the mean - Simulation
Samples drawn from a Normal distribution
}
