% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consens.R
\name{consens}
\alias{consens}
\title{Compute consensus for ordinal data sets}
\usage{
consens(V, tbl = FALSE)
}
\arguments{
\item{tbl}{Logical TRUE/FALSE: Table data or Raw data (default)}

\item{x}{a ordinal data vector}
}
\value{
consensus index (according to Tatsle & Wierman (2007))
}
\description{
Compute consensus for ordinal data sets
}
\examples{
cons <- consens(sample(1:7,200,replace=TRUE),tbl=FALSE)
}
