% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorComb.R
\name{CorComb}
\alias{CorComb}
\title{Combinaison de  k corrélations}
\usage{
CorComb(r, n)
}
\arguments{
\item{r}{= vecteur des Corrélations observées}

\item{n}{= vecteur des Tailles d'échantillons}
}
\description{
Combinaison de  k corrélations
}
