% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTest.R
\name{CorTest}
\alias{CorTest}
\title{Test d'hypothèse sur k corrélations}
\usage{
CorTest(r, n)
}
\arguments{
\item{r}{= vecteur des Corrélations observées}

\item{n}{= vecteur des Tailles d'échantillons}
}
\description{
Test d'hypothèse sur k corrélations
}
