% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiceRoll.R
\name{DiceRoll}
\alias{DiceRoll}
\title{Dice Roll}
\usage{
DiceRoll(n = 1, p = rep(1, 6))
}
\arguments{
\item{n}{number of Dice to roll (integer value, default = 1)}

\item{p}{vector of probabilities of a 'success' (6 probs), default = 6X(1/6)
if it is a single value, the probabilities are equals}
}
\value{
Value of each die
}
\description{
Dice Roll
}
