% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProbDistributions.R, R/exp_area.R
\name{exp_area}
\alias{exp_area}
\title{Adapted from: https://r-coder.com/normal-distribution-r
Plot Exponential distribution}
\usage{
exp_area(rate = 1, lb, ub, acolor = "lightgray", ...)

exp_area(rate = 1, lb, ub, acolor = "lightgray", ...)
}
\arguments{
\item{rate:}{lambda parameter}

\item{lb:}{lower bound of the area}

\item{ub:}{upper bound of the area}

\item{acolor:}{color of the area
...: additional arguments to be passed to the lines function}
}
\description{
Arguments:
rate: lambda parameter
lb: lower bound of the area
ub: upper bound of the area
acolor: color of the area
...: additional arguments to be passed to the lines function
}
