% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exper.R
\name{exper}
\alias{exper}
\title{Run random experiment}
\usage{
exper(x, nrep = 5000, n = 25, nconf = 0.95)
}
\arguments{
\item{x}{Population data}

\item{nrep}{scalar value: number of samples to draw (default = 5000)}

\item{n}{scalar value: sample size (default = 25)}

\item{nconf}{scalar value: Confidence level}
}
\value{
vector containing Population mean, standard error, Percentiles (alpha/2, 1-alpha/2)
}
\description{
Run random experiment
}
\examples{
x1 <- rnorm(50000,100,15)
res1 <- exper(x1,nrep=10000,n=15)
x2 <- rbinom(50000,10,0.7)
res2 <- exper(x2,nrep=10000,n=25,nconf=0.99)
}
