% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cells.R
\name{format_cells}
\alias{format_cells}
\title{Format cells of a data frame
From https://stackoverflow.com/questions/28166168/how-to-change-fontface-bold-italics-for-a-cell-in-a-kable-table-in-rmarkdown}
\usage{
format_cells(df, rows, cols, value = c("italics", "bold", "strikethrough"))
}
\arguments{
\item{df}{data frame}

\item{rows}{row(s) of the cell(s) to format}

\item{cols}{column(s) of the cell(s) to format}

\item{value}{format to apply: one of "italics", "bold", "strikethrough"}
}
\description{
Format cells of a data frame
From https://stackoverflow.com/questions/28166168/how-to-change-fontface-bold-italics-for-a-cell-in-a-kable-table-in-rmarkdown
}
