% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphDist.R
\name{GraphDist}
\alias{GraphDist}
\title{Graph Sampling Distribution}
\usage{
GraphDist(xb, nc = 0.95, parm = "Means")
}
\arguments{
\item{xb}{vector of sampling statistics}

\item{nc}{scalar value: Confidence level (default = 0.95)}

\item{parm}{character value: name of the estimated parameter (default = "Mean")}
}
\description{
Graph Sampling Distribution
}
\examples{
xb <- rnorm(1000,100,2)
GraphDist(xb,nc=0.99,parm="Medians")
}
