% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaitiPopGen.R
\name{HaitiPopGen}
\alias{HaitiPopGen}
\title{Generate simulated Haiti (2019) population: 11.2 millions observations}
\usage{
HaitiPopGen()
}
\value{
Data Frame containing the data structure:
       GENRE: “M” = Masculin “F” = Féminin
       AGE: Âge, en unité d’années
       EDUC: Niveau de scolarité atteint à l’âge de 25 ans - Aucun -
             PréScol - Fond.1 - Fond.2 - Fond.3 - Second - Univ.1 -
             Univ.2 - Univ.3
       DEPARTEMENT: Département de résidence
        "1"  "Artibonite"
        "2"  "Centre"
        "3"  "Grand'Anse"
        "4"  "Nippes"
        "5"  "Nord"
        "6"  "Nord-Est"
        "7"  "Nord-Ouest"
        "8"  "Ouest"
        "9"  "Sud-Est"
        "10" "Sud"

       CondxVie: Indice général de condition de vie, sur une échelle de 0 à 50. Plus l’indice est élevé, plus les conditions de vie de la personne sont bonnes.
       CRIMIDX: Indice de propension à la criminalité et à la corruption, sur une échelle de 0 à 150. Plus l’indice est élevé, plus l’individu a des tendances criminelles.
       HabilCog: Indice d’habiletés cognitives, sur une échelle de 20 à 200. Plus cet indice est élevé, plus les habiletés intellectuelles et cognitives de l’individu sont grandes.
       PrefPres: Préférence pour un candidat donné à la présidentielle de 2022. Cinq candidats sont en liste: CandA - CandE.
       OpinConst: Opinion concernant à un nouveau projet de Constitution:
         “Tout à fait d’accord”
         “D’accord”
         “Indécis”
         “Pas d’accord”
         “Pas du tout d’accord”

L’objet sortant est un ‘DATA.FRAME’ appelé HaitiPop.
}
\description{
Generate simulated Haiti (2019) population: 11.2 millions observations
}
