% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Compute mode of discrete or continuous numeric vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{a numeric vector, discrete or continuous}
}
\value{
mode for continuous data, or mode(s) for discrete data
}
\description{
Compute mode of discrete or continuous numeric vector
}
\examples{
m1 <- Mode(sample(1:7,200,replace=TRUE))
m2 <- Mode(rnorm(200,50,8))
}
