% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModeCat.R
\name{ModeCat}
\alias{ModeCat}
\title{Compute mode & modal frequency of vector (numeric/alphanumeric)}
\usage{
ModeCat(x)
}
\arguments{
\item{x}{a numeric or character vector}
}
\value{
vector containing modal category and modal frequency
}
\description{
Compute mode & modal frequency of vector (numeric/alphanumeric)
}
\examples{
m1 <- ModeCat(sample(1:7,200,replace=TRUE))
m2 <- ModeCat(sample(LETTERS[1:7],200, replace=TRUE))
}
