% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProbDistributions.R, R/normal_area.R
\name{normal_area}
\alias{normal_area}
\title{Adapted from: https://r-coder.com/normal-distribution-r
Plot Normal distribution}
\usage{
normal_area(mean = 0, sd = 1, lb, ub, acolor = "lightgray", ...)

normal_area(mean = 0, sd = 1, lb, ub, acolor = "lightgray", ...)
}
\arguments{
\item{mean:}{mean of the Normal variable}

\item{sd:}{standard deviation of the Normal variable}

\item{lb:}{lower bound of the area}

\item{ub:}{upper bound of the area}

\item{acolor:}{color of the area
...: additional arguments to be passed to lines function}
}
\description{
Arguments:
mean: mean of the Normal variable
sd: standard deviation of the Normal variable
lb: lower bound of the area
ub: upper bound of the area
acolor: color of the area
...: additional arguments to be passed to lines function
}
\examples{
normal_area(mean=100, sd=15, lb=80, ub=125, acolor="red")
}
