% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.n.k.R
\name{perm.n.k}
\alias{perm.n.k}
\title{Compute number of Permutations of n objects taken k at a time}
\usage{
perm.n.k(n, k)
}
\arguments{
\item{n}{Number of items to permute (numeric/alphanumeric)}

\item{k}{Number of items selected from set}
}
\value{
Number of permutations of n objects taken k at a time
}
\description{
Compute number of Permutations of n objects taken k at a time
}
\examples{
p1 <- perm.n.k(n=6,k=2)
p2 <- perm.n.k(10,4)
}
