% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{Permutations of n objects}
\usage{
permute(x)
}
\arguments{
\item{x}{A vector containing n items (numeric/alphanumeric)}
}
\value{
Array of permuted objects, one permutation per line
}
\description{
Generate a full list of permutations of n objects
}
\examples{
p1 <- permute(1:4)
p2 <- permute(letters[1:5])
}
