% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ci.R
\name{plot_ci}
\alias{plot_ci}
\title{Fonction plot_ci: tirée de la librairie Openintro
Plot confidence intervals}
\usage{
plot_ci(lo, hi, m)
}
\arguments{
\item{lo}{a vector of lower bounds for confidence intervals}

\item{hi}{a vector of upper bounds for confidence intervals}

\item{m}{the population mean}
}
\description{
This function takes vectors of upper and lower confidence
interval bounds, as well as population mean, and plots which confidence
intervals intersect the mean
}
