% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProbDistributions.R
\name{plotunif}
\alias{plotunif}
\title{Adapted from: https://r-coder.com/normal-distribution-r
Graph continuous uniform distribution}
\usage{
plotunif(x, min = 0, max = 1, lwd = 1, col = 1, ...)
}
\arguments{
\item{x}{grid of X-axis values (optional)}

\item{min}{lower limit of the distribution (a)}

\item{max}{upper limit of the distribution (b)}

\item{lwd}{line width of the segments of the graph}

\item{col}{color of the segments and points of the graph
...: additional arguments to be passed to the plot function}
}
\description{
Adapted from: https://r-coder.com/normal-distribution-r
Graph continuous uniform distribution
}
