% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotProbDistributions.R
\name{pois_sum}
\alias{pois_sum}
\title{Adapted from: https://r-coder.com/normal-distribution-r
Probability distribution plots}
\usage{
pois_sum(lambda, lb, ub, col = 4, lwd = 1, ...)
}
\arguments{
\item{lambda}{mean}

\item{lb}{lower bound of the sum}

\item{ub}{upper bound of the sum}

\item{col}{color}

\item{lwd}{line width}
}
\description{
Graph Poisson Distribution
}
