% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyUrn.R
\name{polyUrn}
\alias{polyUrn}
\title{Urn}
\usage{
polyUrn(n = 20, r = 5, b = 7, g = 10, repl = FALSE)
}
\arguments{
\item{n}{number of balls to draw from urn (integer value, default = 20)}

\item{r}{number of RED balls in urn (default = 5)}

\item{b}{number of BLUE balls in urn (default = 7)}

\item{g}{number of GREEN balls in urn (default = 10)}

\item{repl:}{with replacement (TRUE/FALSE)}
}
\value{
vector of 3 elements: numbers of balls drawn from each color
}
\description{
Urn
}
