% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_tree.R
\name{prob_tree}
\alias{prob_tree}
\title{Probability Tree
Adapted from: https://daranzolin.github.io/2018-01-07-probability-trees}
\usage{
prob_tree(pa, pba, pab)
}
\arguments{
\item{pa}{Probability that A occurs}

\item{pba}{Probability that B occurs, given A occurred}

\item{pab}{Probability that B desn't occur, given that A didn't occur}
}
\description{
Probability Tree
Adapted from: https://daranzolin.github.io/2018-01-07-probability-trees
}
