% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PwrCalc.R
\name{PwrCalc}
\alias{PwrCalc}
\title{Calcul de puissance: Test d'hypothèse sur une moyenne unique
Arguments:}
\usage{
PwrCalc(n = 30, alpha = 0.05, H0 = 0, H1 = 2, sig = 1, lat = 1)
}
\arguments{
\item{n}{scalaire: taille de l'échantillon (défaut: 30)}

\item{alpha}{scalaire: niveau de signification (défaut: 0.05)}

\item{H0}{scalaire: valeur postulée sous H0 (défaut:0)}

\item{H1}{scalaire: valeur proposée sous H1 (défaut: 2)}

\item{sig}{scalaire: Écart-Type de la population (défaut: 1)}

\item{lat}{scalaire [1, 2]: spécification d'un test uni- ou bilatéral
Sortie:  puissance}
}
\description{
Calcul de puissance: Test d'hypothèse sur une moyenne unique
Arguments:
}
