% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2z.R
\name{r2z}
\alias{r2z}
\title{Fisher Transformation: r to z
Argument: Correlation
Output: Fisher's z}
\usage{
r2z(r)
}
\arguments{
\item{r}{= correlation}
}
\description{
Fisher Transformation: r to z
Argument: Correlation
Output: Fisher's z
}
