% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_CI.R
\name{R_CI}
\alias{R_CI}
\title{Intervalle de confiance d'une corrélation - Simulation
Échantillons tirés d'une distribution normale}
\usage{
R_CI(r = 0.36, n = 50, nrep = 1e+05, conf = 95)
}
\arguments{
\item{r}{= Corrélation au niveau de la population}

\item{n}{= Taille de l'échantillon}

\item{nrep}{= Nombre de ré-échantillonnages}

\item{nconf}{= Niveau de confiance}
}
\description{
Intervalle de confiance d'une corrélation - Simulation
Échantillons tirés d'une distribution normale
}
