% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandUnif.R
\name{RandUnif}
\alias{RandUnif}
\title{Sample from a uniform distribution, with random lower and upper bounds}
\usage{
RandUnif(n = 50, a = c(10, 20), b = c(25, 50))
}
\arguments{
\item{n}{scalar value: Number of observations to draw (default = 50)}

\item{a}{vector containing limits of lower bound of distribution}

\item{b}{vector containing limits of upper bound of distribution}
}
\value{
vector of observations, quantitative values
}
\description{
Sample from a uniform distribution, with random lower and upper bounds
}
\examples{
x1 <- RandUnif() # use defaults
x2 <- RandUnif(25,c(1,5),c(10,15)) # draw 25 observations within
     lower limit set randomly from 1-5, and the upper limit set
     randomly from 10-15
}
