% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegSim.R
\name{RegSim}
\alias{RegSim}
\title{Simple linear regression simulation}
\usage{
RegSim(nrep = 10000, n = 50, xb = 20, sx = 5, yb = 40, sy = 10, r = 0.6)
}
\arguments{
\item{nrep}{scalar value: number of samples (default=10000)}

\item{n}{scalar value: sample size (default = 50)}

\item{xb}{scalar value: mean of independent variable (default = 20)}

\item{sx}{scalar value: standard deviation of dependent variable (default=10)}

\item{yb}{scalar value: mean of dependent variable (default = 40)}

\item{r}{scalar value: correlation coefficient  (default=0.6)}
}
\value{
list containing correlations, intercepts and slopes
}
\description{
Simple linear regression simulation
}
\examples{
 RegSim()
}
