% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampDistrSim.R
\name{SampDistr}
\alias{SampDistr}
\title{This function draws any number of samples of size n from a population vector.
Parametera are:
	- popdata = vector of population values
	- N_Sample = number of sample to draw from the population data (default = 1000)
	- n = sample size (default=25)
Output is a data frame containing:
	- Samp_Means=means: sample means
	- Samp_cil=cil: Sample confidence interval lower limits
	- Samp_ciu=ciu,: Sample confidence interval upper limits
	- Samp_Median=medians: Sample medians
	- Samp_Var=variances: Sample variances
	- Samp_SD=stdeviations: Sample Standard Deviations
	- Samp_MAD=mean_abs_dev: Sample Mean Absolute Deviations
		- Std_Error=std.err.mean: Sample Standard Error of the Means
	- Samp_Skewness=skew: Sample Skewness
		- Samp_Kurtosis=kurtos: Sample Kurtosis
	- Samp_JB=jb: Sample Jarque-Bera Statistics
	- Samp_Size=samp.size: Sample size
	- popdata: Dataset from which samples have been drawn}
\usage{
SampDistr(popdata, N_Sample = 1000, n = 25)
}
\description{
This function draws any number of samples of size n from a population vector.
Parametera are:
	- popdata = vector of population values
	- N_Sample = number of sample to draw from the population data (default = 1000)
	- n = sample size (default=25)
Output is a data frame containing:
	- Samp_Means=means: sample means
	- Samp_cil=cil: Sample confidence interval lower limits
	- Samp_ciu=ciu,: Sample confidence interval upper limits
	- Samp_Median=medians: Sample medians
	- Samp_Var=variances: Sample variances
	- Samp_SD=stdeviations: Sample Standard Deviations
	- Samp_MAD=mean_abs_dev: Sample Mean Absolute Deviations
		- Std_Error=std.err.mean: Sample Standard Error of the Means
	- Samp_Skewness=skew: Sample Skewness
		- Samp_Kurtosis=kurtos: Sample Kurtosis
	- Samp_JB=jb: Sample Jarque-Bera Statistics
	- Samp_Size=samp.size: Sample size
	- popdata: Dataset from which samples have been drawn
}
