% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampDistrSim.R
\name{SampleAnalysis}
\alias{SampleAnalysis}
\title{Function to produce sampling distributions
Input consists of the sampling statistics and population data, as
generated from the SampDistr function
Output consists of a table summarizing the sampling distributions and
histograms of each distributions.}
\usage{
SampleAnalysis(DataIN)
}
\description{
Function to produce sampling distributions
Input consists of the sampling statistics and population data, as
generated from the SampDistr function
Output consists of a table summarizing the sampling distributions and
histograms of each distributions.
}
