% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim2x2.R
\name{sim2x2}
\alias{sim2x2}
\title{Simulation de tableaux de contingence 2x2}
\usage{
sim2x2(p1 = 0.1, p2 = 0.45, n1 = 200, N = 10000)
}
\arguments{
\item{p1}{Probabilité de 'succès' (variable RANGÉES) (Défaut: 0.1)}

\item{p2}{Probabilité de 'succès' (variable Colonnes) (Défaut: 0.45)}

\item{n1}{Taille de l'échantillon (défaut: 200)}

\item{N}{Nombre de tableaux à générer

Sortie: Liste comprenant, pour chaque échantillon:
    a,b,c,d: Fréquences observées dans chacune des cellules
    rr, lrr, selnrr,llrr, ulrr: Rapports de risque, Log(Rapports de risque),
           Erreur-standard des rapports de risque, limites inférieures et
           supérieures (95%) de l'intervalle de confiance des RR.
    or,lnor,selnor,llor,ulor: Rapports de Cotes, Log(Rapports de Cotes),
           Erreur-standard des rapports de Cotes, limites inférieures et
           supérieures (95%) de l'intervalle de confiance des RC.}
}
\description{
Simulation de tableaux de contingence 2x2
}
