% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somersD.R
\name{somersD}
\alias{somersD}
\title{somersD: Calcul du coefficient D de Somers}
\usage{
somersD(tbl, conflev = 0.95, digits = 3)
}
\description{
Arguments:
table: Tableau de contingence bidimensionnel
conflev: Niveau de confiance, 0.95 par défaut
digits: Nombre de décimales, 3 par défaut
}
\details{
Sortie: liste comprenant:
          - Tableau contenant pour VD=R et VD=C,
              - D de Somers
              - Erreur Standard de D
              - Limite inférieure de l'IC
              - Limite supérieure de l'IC
          - Nombre de concordances
          - Nombre de discordances
}
\examples{
x <- matrix(c(12, 4, 3, 5, 10, 6, 3, 5, 14),nrow=3, byrow=TRUE)
somersD(x)
}
