% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.Distr.R
\name{t.Distr}
\alias{t.Distr}
\title{Distribution de Student: Représentation graphique des résultats
d'un test d'hypothèse}
\usage{
\method{t}{Distr}(t, df, alpha = 0.05, h1 = 1)
}
\arguments{
\item{t}{valeur du t observé}

\item{df}{Nombre de degrés de liberté}

\item{alpha}{Niveau de signification}

\item{h1}{Directionalité du test:
-1 = p < 0: unilatéral gauche
+1 = p > 0: unilatéral droite
 0 = p != 0: bilatéral}
}
\description{
Distribution de Student: Représentation graphique des résultats
d'un test d'hypothèse
}
