% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_area.R
\name{t_area}
\alias{t_area}
\title{Adapted from: https://r-coder.com/normal-distribution-r
Function t_area(): draw a Student distribution with shaded area}
\usage{
t_area(df = 25, lb, ub, acolor = "lightgray", ...)
}
\arguments{
\item{df}{Degrees of freedom}

\item{lb}{lower bound of the area}

\item{ub}{upper bound of the area}

\item{acolor}{color of the area
...: additional arguments to be passed to lines function}
}
\description{
Adapted from: https://r-coder.com/normal-distribution-r
Function t_area(): draw a Student distribution with shaded area
}
