% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TblFreq.R
\name{TblFreq}
\alias{TblFreq}
\title{Generate a frequency table}
\usage{
TblFreq(x, bins = 20)
}
\arguments{
\item{x}{a numeric vector}

\item{bins}{number of intervals (bins)}
}
\value{
data frame:
     Lower & Upper limits of intervals,
     Absolute frequencies,
     relative frequencies,
     cumulative frequencies,
     Cumulative/Relative frequencies
}
\description{
Generate a frequency table
}
\examples{
TblFreq(rnorm(200,50,8),n_classes(x))
}
