% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testInt.R
\name{testInt}
\alias{testInt}
\title{Verify that x is an integer value}
\usage{
testInt(x)
}
\arguments{
\item{x}{integer value to check: Must be numeric}
}
\value{
TRUE if integer, FALSE if not
}
\description{
Verify that x is an integer value
}
