% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2r.R
\name{z2r}
\alias{z2r}
\title{Fisher Transformation: z to r
Argument: z
Output: r}
\usage{
z2r(z)
}
\arguments{
\item{z}{= Fisher's zcorrelation}
}
\description{
Fisher Transformation: z to r
Argument: z
Output: r
}
