% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_CI.R
\name{Z_CI}
\alias{Z_CI}
\title{Intervalle de confiance: corrélation transformée en Z (Fisher) - Simulation
Échantillons tirés d'une distribution normale}
\usage{
Z_CI(r = 0.36, n = 50, nrep = 1e+05, conf = 95)
}
\arguments{
\item{r}{= Corrélation au niveau de la population}

\item{n}{= Taille de l'échantillon}

\item{nrep}{= Nombre de réplications}

\item{nconf}{= Niveau de confiance}
}
\description{
Intervalle de confiance: corrélation transformée en Z (Fisher) - Simulation
Échantillons tirés d'une distribution normale
}
