% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlackJack.R
\name{Blackjack}
\alias{Blackjack}
\title{Function Blackjack(): Draw random sets of 3 cards and compute sum of values}
\usage{
Blackjack(sampling_number = 30)
}
\description{
Argument:
sampling_number:  number of card sets to draw (default = 30)
}
